// src/enums.ts
enum RecordType {
	EOF = 0x0000,
	REALIZEPALETTE = 0x0035,
	SETPALENTRIES = 0x0037,
	SETBKMODE = 0x0102,
	SETMAPMODE = 0x0103,
	SETROP2 = 0x0104,
	SETRELABS = 0x0105,
	SETPOLYFILLMODE = 0x0106,
	SETSTRETCHBLTMODE = 0x0107,
	SETTEXTCHAREXTRA = 0x0108,
	RESTOREDC = 0x0127,
	RESIZEPALETTE = 0x0139,
	DIBCREATEPATTERNBRUSH = 0x0142,
	SETLAYOUT = 0x0149,
	SETBKCOLOR = 0x0201,
	SETTEXTCOLOR = 0x0209,
	OFFSETVIEWPORTORG = 0x0211,
	LINETO = 0x0213,
	MOVETO = 0x0214,
	OFFSETCLIPRGN = 0x0220,
	FILLREGION = 0x0228,
	SETMAPPERFLAGS = 0x0231,
	SELECTPALETTE = 0x0234,
	POLYGON = 0x0324,
	POLYLINE = 0x0325,
	SETTEXTJUSTIFICATION = 0x020A,
	SETWINDOWORG = 0x020B,
	SETWINDOWEXT = 0x020C,
	SETVIEWPORTORG = 0x020D,
	SETVIEWPORTEXT = 0x020E,
	OFFSETWINDOWORG = 0x020F,
	SCALEWINDOWEXT = 0x0410,
	SCALEVIEWPORTEXT = 0x0412,
	EXCLUDECLIPRECT = 0x0415,
	INTERSECTCLIPRECT = 0x0416,
	ELLIPSE = 0x0418,
	FLOODFILL = 0x0419,
	FRAMEREGION = 0x0429,
	ANIMATEPALETTE = 0x0436,
	TEXTOUT = 0x0521,
	POLYPOLYGON = 0x0538,
	EXTFLOODFILL = 0x0548,
	RECTANGLE = 0x041B,
	SETPIXEL = 0x041F,
	ROUNDRECT = 0x061C,
	PATBLT = 0x061D,
	SAVEDC = 0x001E,
	PIE = 0x081A,
	STRETCHBLT = 0x0B23,
	ESCAPE = 0x0626,
	INVERTREGION = 0x012A,
	PAINTREGION = 0x012B,
	SELECTCLIPREGION = 0x012C,
	SELECTOBJECT = 0x012D,
	SETTEXTALIGN = 0x012E,
	ARC = 0x0817,
	CHORD = 0x0830,
	BITBLT = 0x0922,
	EXTTEXTOUT = 0x0a32,
	SETDIBTODEV = 0x0d33,
	DIBBITBLT = 0x0940,
	DIBSTRETCHBLT = 0x0b41,
	STRETCHDIB = 0x0f43,
	DELETEOBJECT = 0x01f0,
	CREATEPALETTE = 0x00f7,
	CREATEPATTERNBRUSH = 0x01F9,
	CREATEPENINDIRECT = 0x02FA,
	CREATEFONTINDIRECT = 0x02FB,
	CREATEBRUSHINDIRECT = 0x02FC,
	CREATEREGION = 0x06FF
}

enum PenStyle {
   COSMETIC = 0x0000,
   ENDCAP_ROUND = 0x0000,
   JOIN_ROUND = 0x0000,
   SOLID = 0x0000,
   DASH = 0x0001,
   DOT = 0x0002,
   DASHDOT = 0x0003,
   DASHDOTDOT = 0x0004,
   NULL = 0x0005,
   INSIDEFRAME = 0x0006,
   USERSTYLE = 0x0007,
   ALTERNATE = 0x0008,
   ENDCAP_SQUARE = 0x0100,
   ENDCAP_FLAT = 0x0200,
   JOIN_BEVEL = 0x1000,
   JOIN_MITER = 0x2000
};

enum FamilyFont
{
	DONTCARE = 0x0,
	ROMAN = 0x1,
	SWISS = 0x2,
	MODERN = 0x3,
	SCRIPT = 0x4,
	DECORATIVE = 0x5
};

export { RecordType, PenStyle, FamilyFont };